Adding remove function to remove defined factory
Sometimes, we want to remove defined factory. In my case, I need it because when I use mocha, it watches my files and re-runs project tests, it will say that the factory is defined. It is defined in the first iteration, unfortunately when the test re-run, it will throw that error. This commit will help me to remove defined factory if it's already defined so it won't throw any error.